// Proyect-Vet/Proyect-Vet/src/app/interfaces/QueryListEs.ts
export interface querylistEs {
    ID: number;
    MedicoResponsable: string;
    Fecha: Date;
    NombrePropietario: string;
    Direccion: string;
    Telefono: string;
    Correo: string;
    NombreMascota: string;
    FechaNacimiento: Date;
    FechaIngresoMascota: Date;
    Hora: string;
    Vacunas: string;
    MotivoPresentacion: string;
    MotivoConsulta: string;
    InformacionImportanteMascota: string;
}